pyproject.toml
[project]
name = "mnist-jax-demo"
version = "0.1.0"
description = "MNIST CNN training demo using JAX and Flax NNX"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "tensorflow",
    "jax",
    "jaxlib",
    "flax",
    "optax",
    "tensorflow-datasets",
    "orbax-export",
    "matplotlib",
    "importlib-resources>=6.5.2",
    "protobuf<5",
]
